Make WatchBatteryWidget.kt use BaseWidgetProvider.kt
