+++ mobile/src/main/java/com/boswelja/devicemanager/widget/ui/WidgetSettingsActivity.kt
@@ -14,8 +14,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.datastore.preferences.core.booleanPreferencesKey
 import androidx.datastore.preferences.core.intPreferencesKey
 import com.boswelja.devicemanager.R
-import com.boswelja.devicemanager.batterysync.widget.WatchBatteryWidget
 import com.boswelja.devicemanager.common.ui.AppTheme
+import com.boswelja.devicemanager.common.ui.BaseWidgetProvider
 import com.boswelja.devicemanager.common.ui.UpNavigationAppBar
 
 class WidgetSettingsActivity : AppCompatActivity() {
@@ -45,7 +45,7 @@ class WidgetSettingsActivity : AppCompatActivity() {
 
     override fun onStop() {
         super.onStop()
-        WatchBatteryWidget.updateWidgets(this)
+        BaseWidgetProvider.updateWidgets(this)
     }
 
     companion object {
