+++ mobile/src/main/res/layout/widget_watch_battery.xml
@@ -1,41 +1,29 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout
+<RelativeLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:ignore="MergeRootFrame">
+    android:orientation="vertical">
 
     <ImageView
-        android:id="@+id/widget_background"
+        android:id="@+id/battery_indicator"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
+        android:layout_height="0dp"
+        android:layout_above="@+id/battery_indicator_text"
+        android:layout_alignParentTop="true"
+        android:src="@drawable/watch_battery_unknown"
         tools:ignore="ContentDescription" />
 
-    <RelativeLayout
+    <TextView
+        android:id="@+id/battery_indicator_text"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
+        android:layout_height="wrap_content"
+        android:layout_alignParentBottom="true"
+        android:layout_margin="8dp"
+        android:text="@string/battery_sync_disabled"
+        android:textAlignment="center"
+        android:textColor="@color/widgetForeground"
+        android:textSize="20sp" />
 
-        <ImageView
-            android:id="@+id/battery_indicator"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_above="@+id/battery_indicator_text"
-            android:layout_alignParentTop="true"
-            android:src="@drawable/watch_battery_unknown"
-            tools:ignore="ContentDescription" />
-
-        <TextView
-            android:id="@+id/battery_indicator_text"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_alignParentBottom="true"
-            android:layout_margin="8dp"
-            android:text="@string/battery_sync_disabled"
-            android:textAlignment="center"
-            android:textColor="@color/widgetForeground"
-            android:textSize="20sp" />
-
-    </RelativeLayout>
-</FrameLayout>
+</RelativeLayout>
